requirements.txt
playwright==1.44.0
pyyaml==6.0.1
pytz==2024.1
requests==2.32.3
# Windows-Benachrichtigungen (nur unter Windows aktiv)
winotify==1.1.3; sys_platform == "win32"
